requirements.txt
fastapi
fastapi[standard]
pydantic_settings
sqlmodel
alembic
passlib
bcrypt